chore: extract functions to repo utils module
